make&remove

+++ lab3/zad3.c
@@ -3,7 +3,7 @@
 #include <float.h>
 int main()
 {
-  short a=LONG_MAX; // drukuje "-1" = błąd (short jest mniejszy niż long)
+  short a=(short)LONG_MAX; // drukuje "-1" = błąd (short jest mniejszy niż long)
   a=(short)INT_MAX; // drukuje błąd (--||--)
   int b=(int)FLT_MAX; // drukuje maksimum dla typu float
   b=(int)DBL_MAX; // drukuje maksimum dla double
